+++ packages/client/src/rpc/modules/engine/index.ts
@@ -1,1 +1,2 @@
 export * from './engine'
+export * from './types'
